complete restaurant detail page
