+++ accessible_restaurant/static/accessible_restaurant/browse_style.css
@@ -3156,3 +3156,9 @@
 			}
 
 		}
+
+div.pagebar {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
