+++ accessible_restaurant/templates/accounts/register.html
@@ -21,7 +21,8 @@
             </div>
 
             <div>
-                <h3><a href="{% url 'accessible_restaurant:index' %}">Just Browsing</a></h3>
+<!--                <h3><a href="{% url 'accessible_restaurant:index' %}">Just Browsing</a></h3>-->
+                <h3><a href="{% url 'accessible_restaurant:browse' 0 %}">Just Browsing</a></h3>
                 <p>Search</p>
             </div>
