+++ accessible_restaurant/templates/restaurants/browse.html
@@ -2,7 +2,7 @@
 
 {% block staticlink %}
     {% load static %}
-    <link rel="stylesheet" type="text/css" href="{% static 'accessible_restaurant/browse_style.css' %}">
+    <link rel="stylesheet" type="text/css" href="{% static 'accessible_restaurant/detail_style.css' %}">
 	<script src="https://kit.fontawesome.com/6a247440d5.js" crossorigin="anonymous"></script>
 {% endblock %}
 
@@ -17,7 +17,7 @@
                 <h2><strong>Best Restaurants</strong></h2>
             </header>
             <div class="row 200%">
-                <div class="12u 12u$(medium)">
+                <div class="12u 12u$(medium)" style="align-content:center;">
                     <form method="post" action="#">
                         <div class="row uniform">
                             <div class="9u 12u$(small)">
@@ -29,15 +29,54 @@
                         </div>
                     </form>
 
-                    <div class="box">
-                        <div class="4u 12u$(medium)">
-                            {% for restaurant in restaurants %}
-                                {{ restaurant.img_url }}
-                            {% endfor %}
-                        </div>
-                        <div class="8u 12u$(medium)">
 
-                        </div>
+                    {% for restaurant in restaurants %}
+                        <a href = '{% url "accessible_restaurant:detail" restaurant.business_id %}'>
+                            <div class="box">
+                                <div class="row">
+                                    <div class="4u 12u$(medium)">
+                                        <span class="image fit"><img src="{{ restaurant.img_url }}" alt="" /></span>
+                                    </div>
+                                    <div class="8u 12u$(medium)">
+                                        <header>
+                                            <ul class="actions">
+                                                <li><h2><strong>{{ restaurant.name }}</strong></h2></li>
+                                                {% if restaurant.compliant %}
+                                                    <li><h4><i class="icon far fa-check-circle"></i>compliant</h4></li>
+                                                {% else %}
+                                                    <li><h4><i class="icon far fa-times-circle"></i>compliant</h4></li>
+                                                {% endif %}
+                                            </ul>
+                                            <ul class="icons">
+                                                {% for _ in restaurant.full %}
+                                                    <li><i class="icon fas fa-star"></i></li>
+                                                {% endfor %}
+                                                {% for _ in restaurant.half %}
+                                                    <li><i class="icon fas fa-star-half-alt"></i></li>
+                                                {% endfor %}
+                                                {% for _ in restaurant.null %}
+                                                    <li><i class="icon far fa-star"></i></li>
+                                                {% endfor %}
+                                            </ul>
+                                        </header>
+                                        <ul class="alt">
+                                            <li><i class="icon fas fa-phone-square-alt"></i>{{ restaurant.phone }}</li>
+                                            <li><i class="icon fas fa-map-marker-alt"></i>{{ restaurant.address }}, {{ restaurant.city }}, {{ restaurant.zip_code }}</li>
+                                        </ul>
+                                    </div>
+                                </div>
+                            </div>
+                        </a>
+                    {% endfor %}
+
+                    <div class="row">
+                        <ul class="icons">
+                            <a href="0">First Page</a>
+                            {% for pnum in page_range %}
+                                <li><a href="{{ pnum }}">{{ pnum|add:"1" }}</a></li>
+                            {% endfor %}
+                            <a href="{{ total_page }}">Last Page</a>
+                        </ul>
                     </div>
                 </div>
             </div>
