+++ accessible_restaurant/templates/restaurants/detail.html
@@ -4,6 +4,7 @@
     {% load static %}
     <link rel="stylesheet" type="text/css" href="{% static 'accessible_restaurant/detail_style.css' %}">
 	<script src="https://kit.fontawesome.com/6a247440d5.js" crossorigin="anonymous"></script>
+	<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
 {% endblock %}
 
 {% block title %} Restaurant Detail Page {% endblock %}
@@ -21,7 +22,7 @@
 							{% if restaurant.compliant %}
 								<li><h4><i class="icon far fa-check-circle"></i>compliant</h4></li>
 							{% else %}
-								<li><h4><i class="far fa-times-circle"></i>compliant</h4></li>
+								<li><h4><i class="icon far fa-times-circle"></i>compliant</h4></li>
 							{% endif %}
 						</ul>
 						<ul class="icons">
@@ -29,12 +30,12 @@
 								<li><i class="icon fas fa-star"></i></li>
 							{% endfor %}
 							{% for _ in half %}
-								<li><i class="icon fas a-star-half-alt"></i></li>
+								<li><i class="icon fas fa-star-half-alt"></i></li>
 							{% endfor %}
 							{% for _ in null %}
 								<li><i class="icon far fa-star"></i></li>
 							{% endfor %}
-							<li><h4>115 Reviews</h4></li>
+							<li><h4>{{ restaurant_data.review_count}} Reviews</h4></li>
 						</ul>
 						<ul class="actions">
 							<li><a href="#" class="button special"><i class="far fa-edit"></i>Write a review</a></li>
@@ -90,7 +91,7 @@
 									<span class="image fit"><img src="{{ review.user.image_url }}" alt="" /></span>
 								</div>
 								<div class="row">
-									<h4>{{ review.user.name }}</h4>
+									<h4><a href="{{ review.user.profile_url }}">{{ review.user.name }}</a></h4>
 								</div>
 							</div>
 							<div class="9u 12u$(medium)">
@@ -100,7 +101,7 @@
 											<li><i class="icon fas fa-star"></i></li>
 										{% endfor %}
 										{% for _ in review.half %}
-											<li><i class="icon fas a-star-half-alt"></i></li>
+											<li><i class="icon fas fa-star-half-alt"></i></li>
 										{% endfor %}
 										{% for _ in review.null %}
 											<li><i class="icon far fa-star"></i></li>
@@ -122,6 +123,11 @@
 				<div class="4u$ 12u$(medium)">
 					<h3>Open Hours</h3>
 					<ul>
+						{% if is_open_now %}
+							<li class="no-bullets"><i class="icon fas fa-check-circle"></i>Open Now</li>
+						{% else %}
+							<li class="no-bullets"><i class="icon fas fa-times-circle"></i>Closed</li>
+						{% endif %}
 						{% for hour_data in hours %}
 							<li>{{ hour_data.weekday }} &emsp; {{ hour_data.start }}-{{ hour_data.end }}</li>
 						{% endfor %}
@@ -132,9 +138,51 @@
 						<li><i class="icon fas fa-map-marker-alt"></i>{{ restaurant_data.location.address1 }}, {{ restaurant_data.location.city }}, {{ restaurant_data.location.state }} {{ restaurant_data.location.zip_code }}</li>
 					</ul>
 					<h3>Pictures</h3>
-					<span class="image fit"><img src="{{ restaurant.image_url }}" alt="" /></span>
+						<!-- Slideshow -->
+						<div class="w3-container">
+							{% for image in restaurant_data.photos %}
+							<div class="w3-display-container mySlides">
+							  	<img src="{{ image }}" style="width:100%">
+							</div>
+							{% endfor %}
+							<!-- Slideshow next/previous buttons -->
+							<div class="w3-container w3-green w3-padding w3-large">
+								<div class="w3-left" onclick="plusDivs(-1)"><i class="fa fa-arrow-circle-left w3-hover-text-teal"></i></div>
+								<div class="w3-right" onclick="plusDivs(1)"><i class="fa fa-arrow-circle-right w3-hover-text-teal"></i></div>
+							</div>
+						</div>
+<!--					<span class="image fit"><img src="{{ restaurant.image_url }}" alt="" /></span>-->
 				</div>
 			</div>
 		</div>
 	</section>
+	<script>
+		// Slideshow
+		var slideIndex = 1;
+		showDivs(slideIndex);
+
+		function plusDivs(n) {
+		  showDivs(slideIndex += n);
+		}
+
+		function currentDiv(n) {
+		  showDivs(slideIndex = n);
+		}
+
+		function showDivs(n) {
+		  var i;
+		  var x = document.getElementsByClassName("mySlides");
+		  var dots = document.getElementsByClassName("demodots");
+		  if (n > x.length) {slideIndex = 1}
+		  if (n < 1) {slideIndex = x.length} ;
+		  for (i = 0; i < x.length; i++) {
+			x[i].style.display = "none";
+		  }
+		  for (i = 0; i < dots.length; i++) {
+			dots[i].className = dots[i].className.replace(" w3-white", "");
+		  }
+		  x[slideIndex-1].style.display = "block";
+		  dots[slideIndex-1].className += " w3-white";
+		}
+	</script>
 {% endblock %}
